image_augmentation/remove_this/LANCE/requirements.txt
tqdm
autoflake
matplotlib
seaborn
scipy
torch>=2.0.0
timm==0.4.12
sentencepiece
loguru
opencv-python
diffusers==0.14.0
transformers==4.26.1
pillow==9.0.1
accelerate
notebook
ftfy
ipywidgets
clip @ git+https://github.com/openai/CLIP
nltk
scikit-image
openai
salesforce-lavis
lightning @ git+https://github.com/Lightning-AI/lightning@master
numpy
jsonargparse[signatures]
bitsandbytes
datasets 
zstandard
pydantic >=1.7.4, <2.0.0